refactor: rename ImpressumComponent to ImprintComponent and update related files
